requirements.txt
python-dotenv
torch
torchvision
torchaudio
accelerate
pymupdf
transformers
qwen_vl_utils
bitsandbytes
flash-attn